Refactoring and assign patients fix
EditAppointment was renamed to UpdateAppointment.
The field appointment_id was renamed to id in proto definitions.
Database related code extracted to db.go.
Replaced int64 with int32 in the database.
Closes #6.
